test: fix test_test_util and add tests for global_entry_sink
