+++ metrique-writer-core/src/global.rs
@@ -250,6 +250,18 @@ impl<Q: AttachGlobalEntrySink> GlobalEntrySink for Q {
     }
 }
 
+// We want to test that the doctests our macro generates are runnable. Unfortunately,
+// this is the only way I know of of testing doctests that are generated by macros.
+// Annoying.
+#[cfg(doc)]
+#[doc(hidden)]
+#[allow(unused)]
+mod _doctest {
+    use metrique_writer::sink::global_entry_sink;
+
+    global_entry_sink! { MyEntrySink }
+}
+
 /// Define a new global [`AttachGlobalEntrySink`] that can be referenced by type name in all threads.
 ///
 /// # Usage
@@ -441,7 +453,7 @@ macro_rules! global_entry_sink {
                         /// # Example
                         /// ```rust
                         /// # use metrique_writer::sink::global_entry_sink;
-                        /// # use metrique_writer::test_util::test_entry_sink;
+                        /// # use metrique_writer::test_util::{test_entry_sink, TestEntrySink};
                         /// # global_entry_sink! { TestSink }
                         /// let TestEntrySink { inspector, sink } = test_entry_sink();
                         ///
